Refactor graph handling and introduce ComputationGraph
The computation graph encapsulates the final set of modules executed
by MoMEMta, after sorting and pruning of the modules declared in the
configuration file.

This refactoring now allows us to recompute the real number of dimensions
needed for the integration and update cuba InputTags on-the-fly to ensure
continuous indexing.

Use ComputationGraphBuilder to create a new instance of ComputationGraph

Unit tests have been added, covering most of the common scenario.

Fixes #78